Big Sur update (#9)
Update the documentation to reflect the recent run against a Big Sur MacBook Pro.

Add a `Brewfile` to install items from brew formulae, casks, and the Apple Store. Remove the items from the `setup-my-mac.sh` script and have it run `brew bundle` instead.

Update `init.lua` to reflect the monitor setup is my home office and not tied to a particular company.